requirements.txt
pandas
transformers
sentence-transformers
requests
torch
numpy
faiss-cpu==1.10.0
